requirements.txt
streamlit
transformers
sentence-transformers
faiss-cpu
pymupdf
torch
PyPDF2
numpy
elevenlabs 